Added downloading interview transcript
PT - [Finishes #70416892]
Added a new helper in the transcript service. Deals with fetching the
associated transcript file. From omeka QA_HAS_TRANSCRIPT points to an
expression, but the way I had set it up in the frontend is it just
points directly to the transcript file. So now it just fetches whatever
is correct.